services/rag-ingestor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.0
pydantic==2.5.0
pydantic-settings==2.1.0
tenacity==8.2.3
python-multipart==0.0.6
